[refactor/#59]: utils 디렉토리 생성 및 유틸 함수 분리
