+++ app/study/page.tsx
@@ -2,7 +2,7 @@ import Timer from '@/components/study/Timer';
 import { Container } from '@radix-ui/themes';
 import styles from './page.module.css';
 import SubjectSelectButton from '@/components/study/SubjectSelectButton';
-import { getTimer, postTimer } from '@/api/study';
+import { getTimer } from '@/api/study';
 
 export default async function Study() {
   const timerResponse = await getTimer();
